Update deeplearning_examplerun.R
Dear authors, 
a big thank you for this awesome tool. 
i point out a possible mistake: as.factor generated an error related to format, 
but using the built-in function h2o.asfactor worked for me. I hope I did not miss
something that was loaded previously and made as.factor implicitely referring to
the other function. 
Best, 
david